==> 직장인을_위한_데이터분석_실무_with_파이썬/우리동네_스타벅스/1_crawlingStarbucksList.ipynb <==
이 장에서는 스타벅스의 입지 전략에 대한 가서을 세우고 증명해보는 과정을 갖겠습니다.
1. 거주 인구가 많은 지역에 스타벅스 매자이 많이 입지해 있을 것이다.
2. 직장인이 많은 지역에 스타벅스 매장이 많이 입지해 있을 것이다.

실제로 입지 분석에는 매우 어렵고 복잡한 데이터 분석 기법들이 필요하지만,  
이 책에서는 쉽게 수집할 수 있는 데이터들을 이용해 시각화하는 것만으로도 주요한 인사이트를 얻을 수 있다는 것을 보여드리고자 합니다.

### 5.1 데이터 수집
1. 서울시 내에 출점한 스타벅스들의 위치를 파악해야 합니다.
2. 두 가지 가설을 검증하기 위해서 인구통계 데이터 수집도 필요

#서울시 내에 스타벅스들의 위치 파악을 위한 필요한 모듈
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
from bs4 import BeautifulSoup
import pandas as pd

#크롬 드라이버를 불러오고 스타벅스 매장 url을 엽니다.
driver = webdriver.Chrome('../chromedriver.exe')
url = 'https://www.starbucks.co.kr/store/store_map.do?disp=locale'
driver.get(url)

#서울 버튼을 클릭합니다.
seoul_btn = 'a.set_sido_cd_btn'
driver.find_element_by_css_selector(seoul_btn).click()

#서울 아래 전체를 클릭합니다.
all_btn = 'a.set_gugun_cd_btn'
driver.find_element_by_css_selector(all_btn).click()

#BeautifulSOup 으로 HTML 파서 만들기
html = driver.page_source
soup = BeautifulSoup(html,'html.parser')

#위치 정보 태그를 파악한 후 모두 찾아오기
starbucks_soup_list = soup.select('li.quickResultLstCon')
print(len(starbucks_soup_list))

#태그 문자열 살펴보기
starbucks_soup_list[0]

태그 문자열을 통해 매장 이름(data-name), 주소(p.result_detailes), 위도(data-lat), 경도(dat-long) 등의 정보를 알 수있ㅅ브니다.

# 스타벅스 매장 정보 샘플 확인
starbucks_store = starbucks_soup_list[0]
name = starbucks_store.select('strong')[0].text.strip()
lat = starbucks_store['data-lat'].strip()
lng = starbucks_store['data-long'].strip()
store_type = starbucks_store.select('i')[0]['class'][0][4:]
address = str(starbucks_store.select('p.result_details')[0]).split('<br/>')[0].split('>')[1]
tel = str(starbucks_store.select('p.result_details')[0]).split('<br/>')[1].split('<')[0]

print(name)
print(lat)
print(lng)
print(store_type)
print(address)
print(tel)


#서울시 스타벅스 매장 목록 데이터 만들기
starbucks_list = []
for starbucks_store in starbucks_soup_list:
    name = starbucks_store.select('strong')[0].text.strip()
    lat = starbucks_store['data-lat'].strip()
    lng = starbucks_store['data-long'].strip()
    store_type = starbucks_store.select('i')[0]['class'][0][4:]
    address = str(starbucks_store.select('p.result_details')[0]).split('<br/>')[0].split('>')[1]
    tel = str(starbucks_store.select('p.result_details')[0]).split('<br/>')[1].split('<')[0]

    starbucks_list.append([name,lat,lng,store_type,address,tel])

starbucks_list[:3]

# pandas 데이터프레임 생성
columns = ['매장명','위도','경도','매장타입','주소','전화번호']
seoul_starbucks_df = pd.DataFrame(starbucks_list,columns=columns)
seoul_starbucks_df.head()

#데이터프레임 요약 정보 확인
seoul_starbucks_df.info()

seoul_starbucks_df.to_excel('./files/0416seoul_starbucks_list.xlsx')
