==> 직장인을_위한_데이터분석_실무_with_파이썬/우리동네_스타벅스/2_seoul_open_api.ipynb <==
#### 5.1.2 서울열린데이터광장의 OPEN API를 활용한 공공데이터 수집

서울열린데이터광장(https://data.seoul.go.kr/) 사이트를 통해 행정구역 내의 다양한 공공데이터를 제공

이번 절에서는 서울열린데이터괒앙에서 OPEN API를 사용해 시군구 목록, 주민등록 인구 통계, 사업체 수 정보를 수집해보겠습니다.

#노트북 환경
import sys
sys.path.append('C:/Users/kkang/새 폴더/')
import OpenApiKey

서울시 행정구역 시군구 정보 서울열린데이터광장에서 검색  
이 데이터는 시군구코드, 한글과 영문, 시군구명, 위도, 경도 등의 데이터로 구성되며 엑셀파일로 다운로드 하거나 OPEN API를 통해 획득 가능합니다.  
API의 url 형식  
'http://openapi.seoul.go.kr:8088/{key}/{type}/{serivce}/{star_index}/{end_index/{objectid}'
key : string 필수 인증키   
type : string 필수 요청파일타입 xml/xmlf/xls/json 타입 지원  
service : string 필수 서비스명 SdeTlSccoSigw  
start_index : string 필수 요청시작위치 숫자입력  
end_index : string 필수 요청종료위치 숫자입력  
objectid : string 선택 순번 정수입력  


이 장에서는 파이썬의 dictionary 타입과 비슷한 json 타입을 이용  


#필요한 모듈 
import requests
import pandas as pd

# 서울열린데ㅣ터광장 API 호출 키
SOAK = OpenApiKey.SOAK

# 시군구 목록을 가져오는 API 호출 URL
url = f'http://openapi.seoul.go.kr:8088/{SOAK}/json/SdeTlSccoSigW/1/25'
print(url)

result_dict = requests.get(url).json()

result_dict

# 시군구 목록 결과 딕셔너리 데이터의 값 가져오기
result_data = result_dict['SdeTlSccoSigW']

print(result_data['list_total_count'])
print(result_data['RESULT'])
print(result_data['row'][0])

#결과 데이터를 pandas의 데이터프레임으로 변환
data_list = result_data['row']
sample_df = pd.DataFrame(data_list)
sample_df.head()

***서울열린데이터광장 OPEN API를 호출하는 함수 작성***

def seoul_open_api_data(url,service):
    data_list = None
    try:
        result_dict = requests.get(url).json()
        result_data = result_dict[service]
        code = result_data['RESULT']['CODE']
        if code =='INFO-000':
            data_list = result_data['row']
    except:
        pass
    return data_list
    
        

#
sgg_df =pd.DataFrame(sample_df , columns= ['SIG_CD','SIG_KOR_NM','LAT','LNG'])
sgg_df

sgg_df.columns=['시군구코드','시군구명','위도','경도']
sgg_df.head()

#엑셀로 저장
sgg_df.to_excel('./files/0527seoul_sgg_list.xlsx',index= False)

sgg_df.info()

***공통함수를 이용해 서울시 시군구별 인구 데이터 가져오기***  
http://openapi.seoul.go.kr:8088/{key}/json/octastatapi419/1/5/gigan/jachigu


key : string 필수 인증키   
type : string 필수 요청파일타입 xml/xmlf/xls/json 타입 지원  
service : string 필수 서비스명 SdeTlSccoSigw  
start_index : string 필수 요청시작위치 숫자입력  
end_index : string 필수 요청종료위치 숫자입력  
gigan : string 선택 기간 조회 연도  
JACHIGU : string 선택 시군구 시군구명  



pop_url =f'http://openapi.seoul.go.kr:8088/{SOAK}/json/octastatapi419/1/26/2019.3-4/'
pop_data_list = seoul_open_api_data(pop_url,'octastatapi419')
sgg_pop_df = pd.DataFrame(pop_data_list)
sgg_pop_df.head()

#필요 없는 데이터 제거
condition = sgg_pop_df['JACHIGU'] != '합계'
sgg_pop_df_selected = sgg_pop_df[condition]
sgg_pop_df_selected.head()

sgg_pop_df_selected.info()

#필요한 columns 만 선택
columns = ['JACHIGU','GYE_1']
sgg_pop_df_final = sgg_pop_df_selected[columns]
sgg_pop_df_final.head()

sgg_pop_df_final.columns = ['시군구명','주민등록인구']
sgg_pop_df_final.head()

sgg_pop_df_final.to_excel('./files/0527sgg_pop.xlsx',index= False)

***공통 함수를 이용해 서울시 시군구별 사업체 데이터 가져오기***  
http://openapi.seoul.go.kr:8088/(인증키)/xml/octastatapi104/1/5/  
key : string 필수 인증키   
type : string 필수 요청파일타입 xml/xmlf/xls/json 타입 지원  
service : string 필수 서비스명 SdeTlSccoSigw  
start_index : string 필수 요청시작위치 숫자입력  
end_index : string 필수 요청종료위치 숫자입력  
gigan : string 선택 기간 조회 연도  
JACHIGU : string 선택 시군구 시군구명 

#공통 함수를 이용한 서울시 동별 사업체현황 통계 API 현황
biz_url = f'http://openapi.seoul.go.kr:8088/{SOAK}/json/octastatapi104/1/450/'
biz_data_list = seoul_open_api_data(biz_url,'octastatapi104')

sgg_biz_df = pd.DataFrame(biz_data_list)
sgg_biz_df.head(10)

#시군구별 사업체 현황 데이터 추출
condition = sgg_biz_df['DONG'] == '소계'
sgg_biz_df_selected = sgg_biz_df[condition]
sgg_biz_df_selected.head(10)

#필요없는 칼럼 제거
columns = ['JACHIGU','GYE','SAEOPCHESU_1']
sgg_biz_final = sgg_biz_df_selected[columns]
sgg_biz_final.columns =['시군구명','종사자수','사업체수']
sgg_biz_final.head()

#데이터 프레임 인덱스 초기화
sgg_biz_final = sgg_biz_final.reset_index(drop= True)
sgg_biz_final.head()

#엑셀로 저장
sgg_biz_final.to_excel('./files/0527sgg_biz_df.xlsx',index =False)

### 5.2 데이터 전처리
