==> 직장인을_위한_데이터분석_실무_with_파이썬/우리동네_스타벅스/3_preprocessing.ipynb <==
### 5.2 데이터 전처리

두 개 이상의 데이터를 병합하기 위해서는 공통된 기준값이 필요하며, 이를 키(Key0라고 합니다.  
이 예제에서는 '시군구명'이 키 값이 되며, 이를 기준으로 데이터를 병합하겠습니다.

#### 5.2.1 서울시, 스타벅스 매장 목록, 인구, 사업체 데이터에 시군구명, 시군구코드 추가

# 서울시 스타벅스 매장 목록이 담긴 엑셀 파일 불러오기
import pandas as pd
seoul_starbucks = pd.read_excel('./files/0416seoul_starbucks_list.xlsx',header= 0)
seoul_starbucks.head(10)

del seoul_starbucks['Unnamed: 0']

이 데이터 주소 칼럼에 들어있는 시군구명을 추출

#스타벅스 주소 정보에서 시군구명 추출
seoul_starbucks['시군구명'] = ''
for n, i in enumerate(seoul_starbucks['주소'].str.split()):
    seoul_starbucks['시군구명'][n] = i[1]

seoul_starbucks.head(10)

seoul_starbucks.tail(10)

시군구명이 추가된 서울시 스타벅스 목록 데이터는 데이터 시각화와 분석에 활용할 최종 데이터입니다. 

#엑셀로 저장 
seoul_starbucks.to_excel('./files/0416seoul_starbucks_final.xlsx',index=False)


#### 스타벅스 분석 데이터 만들기
스타벅스 매장 리스트에 시군구명 정보를 추가했으므로 지금부터 이 시군구명 정보를 이용해 데이터 시군구별 스타벅스 매장 수 , 주민인구, 사업체 수 통계 데이터를 병합하겠습니다

#시군구 목록 데이터 불러오기
seoul_sgg = pd.read_excel('./files/0527seoul_sgg_list.xlsx')
seoul_sgg.head(10)

데이터를 병합할 경우 기준이 되는 키는 유일한 값이어야 하지만 명칭의 경우 중복된 명칭을 가지는 등 둘 이상의 값을 가지는 경우가 있습니다.  
따라서 일반적으로는 명칭 대신 각 값에 코드를 부여하고 이를 키로 설정해서 데이터를 관리하고 활용합니다.  
전국을 대상으로 분석을 진행할 경우 시군구코드 등 중복 여부를 점검하고 진행하는 것이 좋습니다.

#서울시 스타벅스 매장 목록 데이터 불러오기
seoul_starbucks= pd.read_excel('./files/0416seoul_starbucks_final.xlsx')
seoul_starbucks.head(10)

시군구명을 이용한 시군구별 매장수 파악가능

#시군구별 스타벅스 매장 수 세기
starbucks_sgg = seoul_starbucks.pivot_table(
    index = '시군구명',
    values = '매장명',
    aggfunc='count').rename(columns = {'매장명':'스타벅스_매장수'})
starbucks_sgg.head()

pandas pivot_table:  
index 항목에는 그룹화할 칼럼명 지정  
values 항목에는 결과에 포함할 칼럼명을 지정  
aggfunc 항목에는 개수를 세는 'count'

그 후 rename을 통해 매장명에서 스타벅스_매장수로 이름 변경

seoul_sgg= pd.merge(seoul_sgg, starbucks_sgg, how= 'left', on = '시군구명')
seoul_sgg.head()

pandas merge()  
병합할 데이터셋 2개를 입력한 후  
join의 기준이 될 데이터셋을 how 항목에서 left로 지정  
key 값을 on 항목에서 시군구명으로 지정

계속해서 merge를 통해 엑셀들을 통합

#서울시 시군구별 인구통계 데이터 불러오기
seoul_sgg_pop = pd.read_excel('./files/0527sgg_pop.xlsx')
seoul_sgg_pop.head()

seoul_sgg = pd.merge(seoul_sgg, seoul_sgg_pop, how= 'left', on = '시군구명')
seoul_sgg.head()

#서울시 시군구별 사업체 수 엑셀 불러온 후 합치기
seoul_sgg_biz = pd.read_excel('./files/0527sgg_biz_df.xlsx')
seoul_sgg = pd.merge(seoul_sgg, seoul_sgg_biz, how = 'left',on = '시군구명')
seoul_sgg.head()

서울시 시군구별 사업체 수 통계 데이터를 불러와 인구통계 데이터를 병합했던 방식과 동일한 방식으로 서울시 시군구 목록 데이터에 병합했습니다.  
병합한 결과를 엑셀 파일에 저장하겠습니다.

seoul_sgg.to_excel('./files/final_sgg_stat.xlsx',index= False)
