==> 직장인을_위한_데이터분석_실무_with_파이썬/우리동네_스타벅스/4_visualization.ipynb <==
### 5.3 데이터 시각화
스타벅스 매장 위치를 지도에 시각화해 서울시 내에서 스타벅스 매장이 어떻게 분포돼 있는지, 일반 매장과, 리저브, 드라이브스루 매장 분포에 어떤 차이가 있는지 확인  
그리고 시군구 단위로 집계한 스타벅스 매장수/인구수/사업체수를 시각화해서 이번 장 서두에 세웠던 두 가지 가설에 대한 검증을 진행해 보겠습니다.

#### 5.3.1 스타벅스 매장분포 시각화
스타벅스 매장 데이터 에는 513개의 스타벅스 매장이 포함돼 있습니다.  
이번 절에서는 모든 스타벅스 매장을 지도에 표시해 서울 내 스타벅스 매장이 어떻게 분포돼 있는지 확인해 보겠습니다.

import pandas as pd
import folium 
import json
